+++ demo/index.js
@@ -53,7 +53,7 @@ window.vditor = new Vditor('vditor', {
   _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
   // _lutePath: 'src/js/lute/lute.min.js',
   toolbar,
-  // mode: 'sv',
+  mode: 'sv',
   height: window.innerHeight + 100,
   outline: true,
   debugger: true,