+++ src/ts/ir/index.ts
@@ -1,6 +1,14 @@
 import {Constants} from "../constants";
 import {isCtrl, isFirefox} from "../util/compatibility";
-import {blurEvent, dropEvent, focusEvent, hotkeyEvent, scrollCenter, selectEvent} from "../util/editorCommonEvent";
+import {
+    blurEvent,
+    copyEvent, cutEvent,
+    dropEvent,
+    focusEvent,
+    hotkeyEvent,
+    scrollCenter,
+    selectEvent,
+} from "../util/editorCommonEvent";
 import {paste} from "../util/fixBrowserBehavior";
 import {hasClosestByClassName} from "../util/hasClosest";
 import {
@@ -35,11 +43,14 @@ class IR {
         hotkeyEvent(vditor, this.element);
         selectEvent(vditor, this.element);
         dropEvent(vditor, this.element);
+        copyEvent(vditor, this.element, this.copy);
+        cutEvent(vditor, this.element, this.copy);
     }
-    private copyEvent(event: ClipboardEvent & { target: HTMLElement }, vditor: IVditor) {
+
+    private copy(event: ClipboardEvent, vditor: IVditor) {
         const range = getSelection().getRangeAt(0);
         if (range.toString() === "") {
-          return;
+            return;
         }
         event.stopPropagation();
         event.preventDefault();
@@ -50,16 +61,8 @@ class IR {
         event.clipboardData.setData("text/plain", vditor.lute.VditorIRDOM2Md(tempElement.innerHTML).trim());
         event.clipboardData.setData("text/html", "");
     }
-    private bindEvent(vditor: IVditor) {
-        const copyEvent = this.copyEvent;
-        this.element.addEventListener(
-            "copy",
-            (event: ClipboardEvent & { target: HTMLElement }) => copyEvent(event, vditor));
 
-        this.element.addEventListener("cut", (event: ClipboardEvent & { target: HTMLElement }) => {
-            copyEvent(event, vditor);
-            document.execCommand("delete");
-        });
+    private bindEvent(vditor: IVditor) {
         this.element.addEventListener("paste", (event: ClipboardEvent & { target: HTMLElement }) => {
             paste(vditor, event, {
                 pasteCode: (code: string) => {